Update A15-binary-logging
Updates to the grfc:

- Do not mandate the use of an env variable for GRPC_BINARY_LOG_FILTER
  across languages
- Specifically call out that the log events reflect what is visible to
  an application
- Removed the section about generating a call ID on client side and
  sending it to the server. binary logs should only observe, and
  should not alter the RPC.
- Removed section about logging in c-core, since there are currently
  discussions about moving the implementation to C++ and python
  interceptors.
- pluggability: removed some unnecessary language
- updated the proto

Further discussions:

Should we require that call id be a monotonically increasing number,
and remove the ID generation interface? The "grpc-trace-bin" header
already contains the census trace info, albeit in a custom (non
protobuf) serialization format. If some headers are special and should
never be truncated, we can expose a configuration knob for that.

Should we remove the payload compression info? In java, we don't have
access to the original compressed data, and full stream decompression
makes it difficult to determine if a message was compressed on the wire.